Rename lib file
