+++ functions/lib/azure-storage-helper.js
